server/requirements.txt
amqp==2.2.1
asgi-redis==1.3.0
asgiref==1.1.2
asn1crypto==0.24.0
attrs==17.2.0
autobahn==17.7.1
Automat==0.6.0
Babel==2.5.1
billiard==3.5.0.3
boto==2.48.0
boto3==1.4.7
botocore==1.7.43
celery==4.1.0
certifi==2017.4.17
cffi==1.11.2
channels==1.1.6
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.1
coreschema==0.0.4
cryptography==2.1.4
daphne==1.3.0
Django==1.11.3
django-cors-headers==2.0.2
django-extensions==1.8.1
django-haystack==2.6.1
django-model-utils==3.0.0
django-rest-swagger==2.1.2
djangorestframework==3.6.2
docutils==0.14
ecdsa==0.13
elasticsearch==2.4.1
flower==0.9.2
http-ece==1.0.5
httplib2==0.10.3
hyperlink==17.3.0
idna==2.5
image==1.5.15
incremental==17.5.0
itypes==1.1.0
Jinja2==2.9.6
jmespath==0.9.3
kombu==4.1.0
MarkupSafe==1.0
msgpack-python==0.4.8
oauth2client==4.1.2
olefile==0.44
openapi-codec==1.3.1
Pillow==4.2.1
psycopg2==2.7.3.2
py-vapid==1.3.0
pyasn1==0.3.4
pyasn1-modules==0.1.4
pycparser==2.18
python-dateutil==2.6.1
pytz==2017.2
pywebpush==1.4.0
redis==2.10.5
requests==2.18.1
rsa==3.4.2
s3transfer==0.1.11
simplejson==3.11.1
six==1.10.0
tornado==4.5.2
Twisted==17.5.0
txaio==2.8.1
uritemplate==3.0.0
urllib3==1.21.1
uWSGI==2.0.15
vine==1.1.4
zope.interface==4.4.2
